+++ packages/server/tests/acceptance/rpc_batch1.spec.ts
@@ -20,25 +20,20 @@
 
 // external resources
 import { expect } from 'chai';
-import { BigNumber, ethers } from 'ethers';
+import { ethers } from 'ethers';
 import { AliasAccount } from '../clients/servicesClient';
 import Assertions from '../helpers/assertions';
 import { Utils } from '../helpers/utils';
 import { ContractFunctionParameters } from '@hashgraph/sdk';
-import TokenCreateJson from '../contracts/TokenCreateContract.json';
 
 // local resources
 import parentContractJson from '../contracts/Parent.json';
-import basicContractJson from '../contracts/Basic.json';
-import storageContractJson from '../contracts/Storage.json';
-import reverterContractJson from '../contracts/Reverter.json';
 import logsContractJson from '../contracts/Logs.json';
 import { predefined } from '../../../relay/src/lib/errors/JsonRpcError';
 import { EthImpl } from '@hashgraph/json-rpc-relay/src/lib/eth';
 import constants from '@hashgraph/json-rpc-relay/src/lib/constants';
-import { resourceLimits } from 'worker_threads';
 
-describe('@api RPC Server Acceptance Tests', function () {
+describe('@api-batch-1 RPC Server Acceptance Tests', function () {
     this.timeout(240 * 1000); // 240 seconds
 
     const accounts: AliasAccount[] = [];
@@ -60,20 +55,12 @@ describe('@api RPC Server Acceptance Tests', function () {
     const INCORRECT_CHAIN_ID = 999;
     const GAS_PRICE_TOO_LOW = '0x1';
     const ONE_TINYBAR = ethers.utils.parseUnits('1', 10).toHexString();
-    const ONE_WEIBAR = ethers.utils.parseUnits('1', 18).toHexString();
 
     const NON_EXISTING_ADDRESS = '0x5555555555555555555555555555555555555555';
     const NON_EXISTING_TX_HASH = '0x5555555555555555555555555555555555555555555555555555555555555555';
     const NON_EXISTING_BLOCK_HASH = '0x5555555555555555555555555555555555555555555555555555555555555555';
     const NON_EXISTING_BLOCK_NUMBER = EthImpl.numberTo0x(99999999);
     const NON_EXISTING_INDEX = EthImpl.numberTo0x(999999);
-    const BASIC_CONTRACT_PING_CALL_DATA = '0x5c36b186';
-    const BASIC_CONTRACT_PING_RESULT = '0x0000000000000000000000000000000000000000000000000000000000000001';
-    const EXCHANGE_RATE_FILE_ID = "0.0.112";
-    const EXCHANGE_RATE_FILE_CONTENT_DEFAULT = "0a1008b0ea0110f9bb1b1a0608f0cccf9306121008b0ea0110e9c81a1a060880e9cf9306";
-    const FEE_SCHEDULE_FILE_ID = "0.0.111";
-    const FEE_SCHEDULE_FILE_CONTENT_DEFAULT = "0a280a0a08541a061a04408888340a0a08061a061a0440889d2d0a0a08071a061a0440b0b63c120208011200"; // Eth gas = 853000
-    const FEE_SCHEDULE_FILE_CONTENT_UPDATED = "0a280a0a08541a061a0440a8953a0a0a08061a061a0440889d2d0a0a08071a061a0440b0b63c120208011200"; // Eth gas = 953000
 
     let blockNumberAtStartOfTests = 0;
     let mirrorAccount0AtStartOfTests;
@@ -84,10 +71,9 @@ describe('@api RPC Server Acceptance Tests', function () {
         this.beforeAll(async () => {
             requestId = Utils.generateRequestId();
 
-            accounts[0] = await servicesNode.createAliasAccount(15, null, requestId);
-            accounts[1] = await servicesNode.createAliasAccount(15, null, requestId);
-            accounts[2] = await servicesNode.createAliasAccount(30, null, requestId);
-            accounts[3] = await servicesNode.createAliasAccount(60, relay.provider, requestId);
+            accounts[0] = await servicesNode.createAliasAccount(10, null, requestId);
+            accounts[1] = await servicesNode.createAliasAccount(10, null, requestId);
+            accounts[2] = await servicesNode.createAliasAccount(40, null, requestId);
             contractId = await accounts[0].client.createParentContract(parentContractJson, requestId);
 
             const params = new ContractFunctionParameters().addUint256(1);
@@ -910,900 +896,6 @@ describe('@api RPC Server Acceptance Tests', function () {
             it('should execute "eth_getTransactionByHash" for non-existing transaction and return null', async function () {
                 const res = await relay.call('eth_getTransactionByHash', [NON_EXISTING_TX_HASH], requestId);
                 expect(res).to.be.null;
-            });
-        });
-
-        describe('eth_estimateGas', async function () {
-            it('@release should execute "eth_estimateGas"', async function () {
-                const res = await relay.call('eth_estimateGas', [{}], requestId);
-                expect(res).to.contain('0x');
-                expect(res).to.not.be.equal('0x');
-                expect(res).to.not.be.equal('0x0');
-            });
-            it('@release should execute "eth_estimateGas" for contract call', async function() {
-                const res = await relay.call('eth_estimateGas', [{
-                    to: mirrorContract.evm_address,
-                    data: BASIC_CONTRACT_PING_CALL_DATA
-                }], requestId);
-                expect(res).to.contain('0x');
-                expect(res).to.equal(EthImpl.defaultGas);
-            });
-            it('@release should execute "eth_estimateGas" for existing account', async function() {
-                const res = await relay.call('eth_estimateGas', [{
-                    to: mirrorSecondaryAccount.evm_address
-                }], requestId);
-                expect(res).to.contain('0x');
-                expect(res).to.equal(EthImpl.gasTxBaseCost);
-            });
-            it('@release should execute "eth_estimateGas" hollow account creation', async function() {
-                const hollowAccount = ethers.Wallet.createRandom();
-                const res = await relay.call('eth_estimateGas', [{
-                    to: hollowAccount.address
-                }], requestId);
-                expect(res).to.contain('0x');
-                expect(res).to.equal(EthImpl.gasTxHollowAccountCreation);
-            });
-        });
-
-        describe('eth_gasPrice', async function () {
-            it('@release should call eth_gasPrice', async function () {
-                const res = await relay.call('eth_gasPrice', [], requestId);
-                expect(res).to.exist;
-                if (process.env.LOCAL_NODE && process.env.LOCAL_NODE !== 'false') {
-                    expect(res).be.equal(ethers.utils.hexValue(Assertions.defaultGasPrice));
-                }
-                else {
-                    expect(Number(res)).to.be.gt(0);
-                }
-            });
-        })
-
-        describe('eth_getBalance', async function () {
-            let getBalanceContractId;
-            before(async function () {
-                getBalanceContractId = await accounts[0].client.createParentContract(parentContractJson, requestId);
-            })
-
-            it('@release should execute "eth_getBalance" for newly created account with 10 HBAR', async function () {
-                const account = await servicesNode.createAliasAccount(10, null, requestId);
-                const mirrorAccount = await mirrorNode.get(`/accounts/${account.accountId}`, requestId);
-
-                const res = await relay.call('eth_getBalance', ['0x' + account.address, 'latest'], requestId);
-                const balanceInWeiBars = BigNumber.from(mirrorAccount.balance.balance.toString()).mul(constants.TINYBAR_TO_WEIBAR_COEF);
-                // balance for tests changes as accounts are in use. Ensure non zero value
-                expect(res).to.not.be.eq(EthImpl.zeroHex);
-            });
-
-            it('should execute "eth_getBalance" for non-existing address', async function () {
-                const res = await relay.call('eth_getBalance', [NON_EXISTING_ADDRESS, 'latest'], requestId);
-                expect(res).to.eq('0x0');
-            });
-
-            it('@release should execute "eth_getBalance" for contract', async function () {
-                const res = await relay.call('eth_getBalance', [Utils.idToEvmAddress(getBalanceContractId.toString()), 'latest'], requestId);
-                expect(res).to.eq(ethers.utils.hexValue(ONE_WEIBAR));
-            });
-
-            it('@release should execute "eth_getBalance" for contract with id converted to evm_address', async function () {
-                const res = await relay.call('eth_getBalance', [Utils.idToEvmAddress(getBalanceContractId.toString()), 'latest'], requestId);
-                expect(res).to.eq(ethers.utils.hexValue(ONE_WEIBAR));
-            });
-
-            it('@release should execute "eth_getBalance" with latest block number', async function () {
-                const latestBlock = (await mirrorNode.get(`/blocks?limit=1&order=desc`, requestId)).blocks[0];
-                const res = await relay.call('eth_getBalance', [Utils.idToEvmAddress(getBalanceContractId.toString()), EthImpl.numberTo0x(latestBlock.number)], requestId);
-                expect(res).to.eq(ethers.utils.hexValue(ONE_WEIBAR));
-            });
-
-            it('@release should execute "eth_getBalance" with one block behind latest block number', async function () {
-                const latestBlock = (await mirrorNode.get(`/blocks?limit=1&order=desc`, requestId)).blocks[0];
-                const res = await relay.call('eth_getBalance', [Utils.idToEvmAddress(getBalanceContractId.toString()), EthImpl.numberTo0x(latestBlock.number - 1)], requestId);
-                expect(res).to.eq(ethers.utils.hexValue(ONE_WEIBAR));
-            });
-
-            it('@release should execute "eth_getBalance" with pending', async function () {
-                const res = await relay.call('eth_getBalance', [Utils.idToEvmAddress(getBalanceContractId.toString()), 'pending'], requestId);
-                expect(res).to.eq(ethers.utils.hexValue(ONE_WEIBAR));
-            });
-
-            it('@release should execute "eth_getBalance" with block number in the last 15 minutes', async function () {
-                const latestBlock = (await mirrorNode.get(`/blocks?limit=1&order=desc`, requestId)).blocks[0];
-                const earlierBlockNumber = latestBlock.number - 2;
-                const res = await relay.call('eth_getBalance', [Utils.idToEvmAddress(getBalanceContractId.toString()), EthImpl.numberTo0x(earlierBlockNumber)], requestId);
-                expect(res).to.eq(ethers.utils.hexValue(ONE_WEIBAR));
-            });
-
-            it('@release should execute "eth_getBalance" with block number in the last 15 minutes for account that has performed contract deploys/calls"', async function () {
-                const res = await relay.call('eth_getBalance', ['0x' + accounts[0].address,  EthImpl.numberTo0x(blockNumberAtStartOfTests)], requestId);
-                const balanceAtBlock = mirrorAccount0AtStartOfTests.balance.balance * constants.TINYBAR_TO_WEIBAR_COEF;
-                expect(res).to.eq(`0x${balanceAtBlock.toString(16)}`);
-            });
-
-            it('@release should correctly execute "eth_getBalance" with block number in the last 15 minutes with several txs around that time', async function () {
-                const initialBalance = await relay.call('eth_getBalance', ['0x' + accounts[0].address, 'latest'], requestId);
-
-                const acc3Nonce = await relay.getAccountNonce('0x' + accounts[3].address);
-                const gasPrice = await relay.gasPrice();
-
-                const transaction = {
-                    value: ONE_TINYBAR,
-                    gasLimit: 50000,
-                    chainId: Number(CHAIN_ID),
-                    to: accounts[0].wallet.address,
-                    nonce: acc3Nonce,
-                    gasPrice: gasPrice,
-                };
-
-                const signedTx1 = await accounts[3].wallet.signTransaction(transaction);
-                const txHash1 = await relay.call('eth_sendRawTransaction', [signedTx1]);
-                await mirrorNode.get(`/contracts/results/${txHash1}`, requestId);
-                const tx1 = await relay.call('eth_getTransactionByHash', [txHash1]);
-                await new Promise(r => setTimeout(r, 4000));
-
-                const signedTx2 = await accounts[3].wallet.signTransaction({...transaction, nonce: acc3Nonce + 1});
-                const txHash2 = await relay.call('eth_sendRawTransaction', [signedTx2]);
-                await mirrorNode.get(`/contracts/results/${txHash2}`, requestId);
-                const tx2 = await relay.call('eth_getTransactionByHash', [txHash2]);
-                await new Promise(r => setTimeout(r, 4000));
-
-                const signedTx3 = await accounts[3].wallet.signTransaction({...transaction, nonce: acc3Nonce + 2});
-                const txHash3 = await relay.call('eth_sendRawTransaction', [signedTx3]);
-                await mirrorNode.get(`/contracts/results/${txHash3}`, requestId);
-                const tx3 = await relay.call('eth_getTransactionByHash', [txHash3]);
-                await new Promise(r => setTimeout(r, 4000));
-
-                const endBalance = await relay.call('eth_getBalance', ['0x' + accounts[0].address, 'latest'], requestId);
-
-                // initialBalance + sum of value of all transactions
-                const manuallyCalculatedBalance = BigNumber.from(initialBalance).add(BigNumber.from(ONE_TINYBAR).mul(3));
-                expect(BigNumber.from(endBalance).toString()).to.eq(manuallyCalculatedBalance.toString());
-
-                // Balance at the block number of tx1 should be initialBalance + the value of tx1
-                const balanceAtTx1Block = await relay.call('eth_getBalance', ['0x' + accounts[0].address, tx1.blockNumber], requestId);
-                const manuallyCalculatedBalanceAtTx1Block = BigNumber.from(initialBalance).add(BigNumber.from(ONE_TINYBAR));
-                expect(BigNumber.from(balanceAtTx1Block).toString()).to.eq(manuallyCalculatedBalanceAtTx1Block.toString());
-            });
-
-        });
-
-        describe('@release Hardcoded RPC Endpoints', () => {
-            let mirrorBlock;
-
-            before(async () => {
-                mirrorBlock = (await mirrorNode.get(`/blocks?block.number=${mirrorContractDetails.block_number}`, requestId)).blocks[0];
-            });
-
-            it('should execute "eth_chainId"', async function () {
-                const res = await relay.call('eth_chainId', [null], requestId);
-                expect(res).to.be.equal(CHAIN_ID);
-            });
-
-            it('should execute "net_listening"', async function () {
-                const res = await relay.call('net_listening', [], requestId);
-                expect(res).to.be.equal('false');
-            });
-
-            it('should execute "net_version"', async function () {
-                const res = await relay.call('net_version', [], requestId);
-                expect(res).to.be.equal(CHAIN_ID);
-            });
-
-            it('should execute "eth_getUncleByBlockHashAndIndex"', async function () {
-                const res = await relay.call('eth_getUncleByBlockHashAndIndex', [mirrorBlock.hash, 0], requestId);
-                expect(res).to.be.null;
-            });
-
-            it('should execute "eth_getUncleByBlockHashAndIndex" for non-existing block hash and index=0', async function () {
-                const res = await relay.call('eth_getUncleByBlockHashAndIndex', [NON_EXISTING_BLOCK_HASH, 0], requestId);
-                expect(res).to.be.null;
-            });
-
-            it('should execute "eth_getUncleByBlockNumberAndIndex"', async function () {
-                const res = await relay.call('eth_getUncleByBlockNumberAndIndex', [mirrorBlock.number, 0], requestId);
-                expect(res).to.be.null;
-            });
-
-            it('should execute "eth_getUncleByBlockNumberAndIndex" for non-existing block number and index=0', async function () {
-                const res = await relay.call('eth_getUncleByBlockNumberAndIndex', [NON_EXISTING_BLOCK_NUMBER, 0], requestId);
-                expect(res).to.be.null;
-            });
-
-            it('should execute "eth_getUncleCountByBlockHash"', async function () {
-                const res = await relay.call('eth_getUncleCountByBlockHash', [], requestId);
-                expect(res).to.be.equal('0x0');
-            });
-
-            it('should execute "eth_getUncleCountByBlockNumber"', async function () {
-                const res = await relay.call('eth_getUncleCountByBlockNumber', [], requestId);
-                expect(res).to.be.equal('0x0');
-            });
-
-            it('should return empty on "eth_accounts"', async function () {
-                const res = await relay.call('eth_accounts', [], requestId);
-                expect(res).to.deep.equal([]);
-            });
-
-            it('should execute "eth_hashrate"', async function () {
-                const res = await relay.call('eth_hashrate', [], requestId);
-                expect(res).to.be.equal('0x0');
-            });
-
-            it('should execute "eth_mining"', async function () {
-                const res = await relay.call('eth_mining', [], requestId);
-                expect(res).to.be.equal(false);
-            });
-
-            it('should execute "eth_submitWork"', async function () {
-                const res = await relay.call('eth_submitWork', [], requestId);
-                expect(res).to.be.equal(false);
-            });
-
-            it('should execute "eth_syncing"', async function () {
-                const res = await relay.call('eth_syncing', [], requestId);
-                expect(res).to.be.equal(false);
-            });
-
-            it('should execute "web3_client_version"', async function () {
-                const res = await relay.call('web3_client_version', [], requestId);
-                expect(res).to.contain('relay/');
-            });
-
-            it('should execute "eth_maxPriorityFeePerGas"', async function () {
-                const res = await relay.call('eth_maxPriorityFeePerGas', [], requestId);
-                expect(res).to.be.equal('0x0');
-            });
-        });
-
-        describe('@release Unsupported RPC Endpoints', () => {
-
-            it('should not support "eth_submitHashrate"', async function () {
-                await relay.callUnsupported('eth_submitHashrate', [], requestId);
-            });
-
-            it('should not support "eth_getWork"', async function () {
-                await relay.callUnsupported('eth_getWork', [], requestId);
-            });
-
-            it('should not support "eth_coinbase"', async function () {
-                await relay.callUnsupported('eth_coinbase', [], requestId);
-            });
-
-            it('should not support "eth_sendTransaction"', async function () {
-                await relay.callUnsupported('eth_sendTransaction', [], requestId);
-            });
-
-            it('should not support "eth_protocolVersion"', async function () {
-                await relay.callUnsupported('eth_protocolVersion', [], requestId);
-            });
-
-            it('should not support "eth_sign"', async function () {
-                await relay.callUnsupported('eth_sign', [], requestId);
-            });
-
-            it('should not support "eth_signTransaction"', async function () {
-                await relay.callUnsupported('eth_signTransaction', [], requestId);
-            });
-        });
-
-        describe('eth_getCode', () => {
-
-            let basicContract;
-            let mainContractAddress: string;
-            let NftHTSTokenContractAddress: string;
-            let redirectBytecode: string;
-
-            async function deploymainContract() {
-                const mainFactory = new ethers.ContractFactory(TokenCreateJson.abi, TokenCreateJson.bytecode, accounts[3].wallet);
-                const mainContract = await mainFactory.deploy({gasLimit: 15000000});
-                const { contractAddress } = await mainContract.deployTransaction.wait();
-
-                return contractAddress;
-            }
-
-            async function createNftHTSToken() {
-                const mainContract = new ethers.Contract(mainContractAddress, TokenCreateJson.abi, accounts[3].wallet);
-                const tx = await mainContract.createNonFungibleTokenPublic(accounts[3].wallet.address, {
-                    value: ethers.BigNumber.from('10000000000000000000'),
-                    gasLimit: 10000000
-                });
-                const { tokenAddress } = (await tx.wait()).events.filter(e => e.event = 'CreatedToken')[0].args;
-
-                return tokenAddress;
-            }
-
-            before(async () => {
-                basicContract = await servicesNode.deployContract(basicContractJson);
-                mainContractAddress = await deploymainContract();
-                NftHTSTokenContractAddress = await createNftHTSToken();
-                // Wait for creation to propagate
-                await mirrorNode.get(`/contracts/${basicContract.contractId}`, requestId);
-            });
-
-            it('should execute "eth_getCode" for hts token', async function () {
-                const tokenAddress = NftHTSTokenContractAddress.slice(2);
-                redirectBytecode = `6080604052348015600f57600080fd5b506000610167905077618dc65e${tokenAddress}600052366000602037600080366018016008845af43d806000803e8160008114605857816000f35b816000fdfea2646970667358221220d8378feed472ba49a0005514ef7087017f707b45fb9bf56bb81bb93ff19a238b64736f6c634300080b0033`
-                const res = await relay.call('eth_getCode', [NftHTSTokenContractAddress, 'latest'], requestId);
-                expect(res).to.equal(redirectBytecode);
-            });
-
-            it('@release should execute "eth_getCode" for contract evm_address', async function () {
-                const evmAddress = basicContract.contractId.toSolidityAddress();
-                const res = await relay.call('eth_getCode', ['0x' + evmAddress, 'latest'], requestId);
-                expect(res).to.eq(basicContractJson.deployedBytecode);
-            });
-
-            it('@release should execute "eth_getCode" for contract with id converted to evm_address', async function () {
-                const evmAddress = Utils.idToEvmAddress(basicContract.contractId.toString());
-                const res = await relay.call('eth_getCode', [evmAddress, 'latest'], requestId);
-                expect(res).to.eq(basicContractJson.deployedBytecode);
-            });
-
-            it('should return 0x0 for non-existing contract on eth_getCode', async function () {
-                const res = await relay.call('eth_getCode', [NON_EXISTING_ADDRESS, 'latest'], requestId);
-                expect(res).to.eq(EthImpl.emptyHex);
-            });
-
-            it('should return 0x0 for account evm_address on eth_getCode', async function () {
-                const evmAddress = Utils.idToEvmAddress(accounts[2].accountId.toString());
-                const res = await relay.call('eth_getCode', [evmAddress, 'latest'], requestId);
-                expect(res).to.eq(EthImpl.emptyHex);
-            });
-
-            it('should return 0x0 for account alias on eth_getCode', async function () {
-                const alias = Utils.idToEvmAddress(accounts[2].accountId.toString());
-                const res = await relay.call('eth_getCode', [alias, 'latest'], requestId);
-                expect(res).to.eq(EthImpl.emptyHex);
-            });
-        });
-
-        describe('eth_call', () => {
-            let basicContract, evmAddress;
-
-            before(async () => {
-                basicContract = await servicesNode.deployContract(basicContractJson);
-                // Wait for creation to propagate
-                await mirrorNode.get(`/contracts/${basicContract.contractId}`, requestId);
-
-                evmAddress = `0x${basicContract.contractId.toSolidityAddress()}`;
-            });
-
-            it('@release should execute "eth_call" request to Basic contract', async function () {
-                const callData = {
-                    from: '0x' + accounts[2].address,
-                    to: evmAddress,
-                    gas: EthImpl.numberTo0x(30000),
-                    data: BASIC_CONTRACT_PING_CALL_DATA
-                };
-
-                const res = await relay.call('eth_call', [callData, 'latest'], requestId);
-                expect(res).to.eq(BASIC_CONTRACT_PING_RESULT);
-            });
-
-            it('should fail "eth_call" request without data field', async function () {
-                const callData = {
-                    from: '0x' + accounts[2].address,
-                    to: evmAddress,
-                    gas: EthImpl.numberTo0x(30000)
-                };
-
-                const res = await relay.call('eth_call', [callData, 'latest'], requestId);
-                expect(res).to.eq('0x'); // confirm no error
-            });
-
-            it('should fail "eth_call" for non-existing contract address', async function () {
-                const callData = {
-                    from: '0x' + accounts[2].address,
-                    to: NON_EXISTING_ADDRESS,
-                    gas: EthImpl.numberTo0x(30000),
-                    data: BASIC_CONTRACT_PING_CALL_DATA
-                };
-
-                await relay.callFailing('eth_call', [callData, 'latest'], predefined.INTERNAL_ERROR(), requestId);
-            });
-
-            it('should execute "eth_call" without from field', async function () {
-                const callData = {
-                    to: evmAddress,
-                    gas: EthImpl.numberTo0x(30000),
-                    data: BASIC_CONTRACT_PING_CALL_DATA
-                };
-
-                const res = await relay.call('eth_call', [callData, 'latest'], requestId);
-                expect(res).to.eq(BASIC_CONTRACT_PING_RESULT);
-            });
-
-            it('should execute "eth_call" without gas field', async function () {
-                const callData = {
-                    from: '0x' + accounts[2].address,
-                    to: evmAddress,
-                    data: BASIC_CONTRACT_PING_CALL_DATA
-                };
-
-                const res = await relay.call('eth_call', [callData, 'latest'], requestId);
-                expect(res).to.eq(BASIC_CONTRACT_PING_RESULT);
-            });
-
-            it('should execute "eth_call" with correct block number', async function () {
-                const callData = {
-                    from: '0x' + accounts[2].address,
-                    to: evmAddress,
-                    data: BASIC_CONTRACT_PING_CALL_DATA
-                };
-                
-                const res = await relay.call('eth_call', [callData, '0x1'], requestId);
-                expect(res).to.eq(BASIC_CONTRACT_PING_RESULT);
-            });
-
-            it('should execute "eth_call" with correct block hash object', async function () {
-                const blockHash = '0xd4e56740f876aef8c010b86a40d5f56745a118d0906a34e69aec8c0db1cb8fa3';
-                const callData = {
-                    from: '0x' + accounts[2].address,
-                    to: evmAddress,
-                    data: BASIC_CONTRACT_PING_CALL_DATA
-                };
-                
-                const res = await relay.call('eth_call', [callData, {'blockHash' : blockHash}], requestId);
-                expect(res).to.eq(BASIC_CONTRACT_PING_RESULT);
-            });
-
-            it('should execute "eth_call" with correct block number object', async function () {
-                const callData = {
-                    from: '0x' + accounts[2].address,
-                    to: evmAddress,
-                    data: BASIC_CONTRACT_PING_CALL_DATA
-                };
-                
-                const res = await relay.call('eth_call', [callData, {'blockNumber' : '0x1'}], requestId);
-                expect(res).to.eq(BASIC_CONTRACT_PING_RESULT);
-            });
-
-            it('should fail to execute "eth_call" with wrong block tag', async function () {
-                const callData = {
-                    from: '0x' + accounts[2].address,
-                    to: evmAddress,
-                    data: BASIC_CONTRACT_PING_CALL_DATA
-                };
-                
-                try {
-                    await relay.call('eth_call', [callData, 'newest'], requestId);
-                    Assertions.expectedError();
-                } catch (error) {
-                    Assertions.jsonRpcError(error,predefined.INVALID_PARAMETER(1, 'Expected 0x prefixed string representing the hash (32 bytes) in object, 0x prefixed hexadecimal block number, or the string "latest", "earliest" or "pending'));
-                }
-            });
-
-            it('should fail to execute "eth_call" with wrong block number', async function () {
-                const callData = {
-                    from: '0x' + accounts[2].address,
-                    to: evmAddress,
-                    data: BASIC_CONTRACT_PING_CALL_DATA
-                };
-                
-                try {
-                    await relay.call('eth_call', [callData, '123'], requestId);
-                    Assertions.expectedError();
-                } catch (error) {
-                    Assertions.jsonRpcError(error,predefined.INVALID_PARAMETER(1, 'Expected 0x prefixed string representing the hash (32 bytes) in object, 0x prefixed hexadecimal block number, or the string "latest", "earliest" or "pending'));
-                }
-            });
-
-            it('should fail to execute "eth_call" with wrong block hash object', async function () {
-                const callData = {
-                    from: '0x' + accounts[2].address,
-                    to: evmAddress,
-                    data: BASIC_CONTRACT_PING_CALL_DATA
-                };
-                
-                try {
-                    await relay.call('eth_call', [callData, {'blockHash' : '0x123'}], requestId);
-                    Assertions.expectedError();
-                } catch (error) {
-
-                    Assertions.jsonRpcError(error,predefined.INVALID_PARAMETER(`'blockHash' for BlockHashObject`, 'Expected 0x prefixed string representing the hash (32 bytes) of a block'));
-                }
-            });
-
-            it('should fail to execute "eth_call" with wrong block number object', async function () {
-                const callData = {
-                    from: '0x' + accounts[2].address,
-                    to: evmAddress,
-                    data: BASIC_CONTRACT_PING_CALL_DATA
-                };
-                
-                try {
-                    await relay.call('eth_call', [callData, {'blockNumber' : '123'}], requestId);
-                    Assertions.expectedError();
-                } catch (error) {
-                    Assertions.jsonRpcError(error,predefined.INVALID_PARAMETER(`'blockNumber' for BlockNumberObject`, 'Expected 0x prefixed hexadecimal block number, or the string "latest", "earliest" or "pending"'));
-                }
-            });
-        });
-
-        // Test state changes with getStorageAt
-        describe('eth_getStorageAt', () => {
-            let storageContract, evmAddress;
-            const STORAGE_CONTRACT_UPDATE = "0x2de4e884";
-            const NEXT_STORAGE_CONTRACT_UPDATE = "0x160D6484";
-
-            this.beforeEach(async () => {
-                storageContract = await servicesNode.deployContract(storageContractJson);
-                // Wait for creation to propagate
-                await mirrorNode.get(`/contracts/${storageContract.contractId}`);
-
-                evmAddress = `0x${storageContract.contractId.toSolidityAddress()}`;
-            });
-
-            it('should execute "eth_getStorageAt" request to get current state changes', async function () {
-                const BEGIN_EXPECTED_STORAGE_VAL = "0x000000000000000000000000000000000000000000000000000000000000000f";
-                const END_EXPECTED_STORAGE_VAL = "0x0000000000000000000000000000000000000000000000000000000000000008";
-
-                const beginStorageVal = await relay.call('eth_getStorageAt', [evmAddress, '0x0000000000000000000000000000000000000000000000000000000000000000', 'latest'], requestId);
-                expect(beginStorageVal).to.eq(BEGIN_EXPECTED_STORAGE_VAL);
-
-                const gasPrice = await relay.gasPrice();
-                const transaction = {
-                    value: 0,
-                    gasLimit: 50000,
-                    chainId: Number(CHAIN_ID),
-                    to: evmAddress,
-                    nonce: await relay.getAccountNonce('0x' + accounts[1].address),
-                    gasPrice: gasPrice,
-                    data: STORAGE_CONTRACT_UPDATE,
-                    maxPriorityFeePerGas: gasPrice,
-                    maxFeePerGas: gasPrice,
-                    type: 2
-                };
-
-                const signedTx = await accounts[1].wallet.signTransaction(transaction);
-                await relay.call('eth_sendRawTransaction', [signedTx], requestId);
-
-                // wait for the transaction to propogate to mirror node
-                await new Promise(r => setTimeout(r, 2000));
-
-                const storageVal = await relay.call('eth_getStorageAt', [evmAddress, '0x0000000000000000000000000000000000000000000000000000000000000000', 'latest'], requestId);
-                expect(storageVal).to.eq(END_EXPECTED_STORAGE_VAL);
-            });
-
-            it('should execute "eth_getStorageAt" request to get current state changes without passing block', async function () {
-                const BEGIN_EXPECTED_STORAGE_VAL = "0x000000000000000000000000000000000000000000000000000000000000000f";
-                const END_EXPECTED_STORAGE_VAL = "0x0000000000000000000000000000000000000000000000000000000000000008";
-
-                const beginStorageVal = await relay.call('eth_getStorageAt', [evmAddress, '0x0000000000000000000000000000000000000000000000000000000000000000'], requestId);
-                expect(beginStorageVal).to.eq(BEGIN_EXPECTED_STORAGE_VAL);
-
-                const gasPrice = await relay.gasPrice();
-                const transaction = {
-                    value: 0,
-                    gasLimit: 50000,
-                    chainId: Number(CHAIN_ID),
-                    to: evmAddress,
-                    nonce: await relay.getAccountNonce('0x' + accounts[1].address),
-                    gasPrice: gasPrice,
-                    data: STORAGE_CONTRACT_UPDATE,
-                    maxPriorityFeePerGas: gasPrice,
-                    maxFeePerGas: gasPrice,
-                    type: 2
-                };
-
-                const signedTx = await accounts[1].wallet.signTransaction(transaction);
-                await relay.call('eth_sendRawTransaction', [signedTx], requestId);
-
-                // wait for the transaction to propogate to mirror node
-                await new Promise(r => setTimeout(r, 2000));
-
-                const storageVal = await relay.call('eth_getStorageAt', [evmAddress, '0x0000000000000000000000000000000000000000000000000000000000000000'], requestId);
-                expect(storageVal).to.eq(END_EXPECTED_STORAGE_VAL);
-            });
-
-            it('should execute "eth_getStorageAt" request to get current state changes with passing specific block', async function () {
-                const EXPECTED_STORAGE_VAL = "0x0000000000000000000000000000000000000000000000000000000000000008";
-
-                const gasPrice = await relay.gasPrice();
-                const transaction = {
-                    value: 0,
-                    gasLimit: 50000,
-                    chainId: Number(CHAIN_ID),
-                    to: evmAddress,
-                    nonce: await relay.getAccountNonce('0x' + accounts[1].address),
-                    gasPrice: gasPrice,
-                    data: STORAGE_CONTRACT_UPDATE,
-                    maxPriorityFeePerGas: gasPrice,
-                    maxFeePerGas: gasPrice,
-                    type: 2
-                };
-
-                const signedTx = await accounts[1].wallet.signTransaction(transaction);
-                const transactionHash = await relay.call('eth_sendRawTransaction', [signedTx], requestId);
-                const blockNumber = await relay.call('eth_getTransactionReceipt', [transactionHash], requestId).blockNumber;
-
-                const transaction1 = {
-                    ...transaction,
-                    nonce: await relay.getAccountNonce('0x' + accounts[1].address),
-                    data: NEXT_STORAGE_CONTRACT_UPDATE,
-                };
-
-                const signedTx1 = await accounts[1].wallet.signTransaction(transaction1);
-                const transactionHash1 = await relay.call('eth_sendRawTransaction', [signedTx1], requestId);
-                await new Promise(r => setTimeout(r, 2000));
-
-                //Get previous state change with specific block number
-                const storageVal = await relay.call('eth_getStorageAt', [evmAddress, '0x0000000000000000000000000000000000000000000000000000000000000000', blockNumber], requestId);
-                expect(storageVal).to.eq(EXPECTED_STORAGE_VAL);
-            });
-        });
-
-        // Only run the following tests against a local node since they only work with the genesis account
-        if (process.env.LOCAL_NODE && process.env.LOCAL_NODE !== 'false') {
-            describe('Gas Price related RPC endpoints', () => {
-                let lastBlockBeforeUpdate;
-                let lastBlockAfterUpdate;
-
-                before(async () => {
-                    await servicesNode.updateFileContent(FEE_SCHEDULE_FILE_ID, FEE_SCHEDULE_FILE_CONTENT_DEFAULT, requestId);
-                    await servicesNode.updateFileContent(EXCHANGE_RATE_FILE_ID, EXCHANGE_RATE_FILE_CONTENT_DEFAULT, requestId);
-                    lastBlockBeforeUpdate = (await mirrorNode.get(`/blocks?limit=1&order=desc`, requestId)).blocks[0];
-                    await new Promise(resolve => setTimeout(resolve, 4000));
-                    await servicesNode.updateFileContent(FEE_SCHEDULE_FILE_ID, FEE_SCHEDULE_FILE_CONTENT_UPDATED, requestId);
-                    await new Promise(resolve => setTimeout(resolve, 4000));
-                    lastBlockAfterUpdate = (await mirrorNode.get(`/blocks?limit=1&order=desc`, requestId)).blocks[0];
-                });
-
-                it('should call eth_feeHistory with updated fees', async function () {
-                    const blockCountNumber = lastBlockAfterUpdate.number - lastBlockBeforeUpdate.number;
-                    const blockCountHex = ethers.utils.hexValue(blockCountNumber);
-                    const datedGasPriceHex = ethers.utils.hexValue(Assertions.datedGasPrice);
-                    const updatedGasPriceHex = ethers.utils.hexValue(Assertions.updatedGasPrice);
-                    const newestBlockNumberHex = ethers.utils.hexValue(lastBlockAfterUpdate.number);
-                    const oldestBlockNumberHex = ethers.utils.hexValue(lastBlockAfterUpdate.number - blockCountNumber + 1);
-
-                    const res = await relay.call('eth_feeHistory', [blockCountHex, newestBlockNumberHex, [0]], requestId);
-
-                    Assertions.feeHistory(res, {
-                        resultCount: blockCountNumber,
-                        oldestBlock: oldestBlockNumberHex,
-                        checkReward: true
-                    });
-                    // We expect all values in the array to be from the mirror node. If there is discrepancy in the blocks, the first value is from the consensus node and it's different from expected.
-                    expect(res.baseFeePerGas[1]).to.equal(datedGasPriceHex);
-                    expect(res.baseFeePerGas[res.baseFeePerGas.length - 2]).to.equal(updatedGasPriceHex);
-                    expect(res.baseFeePerGas[res.baseFeePerGas.length - 1]).to.equal(updatedGasPriceHex);
-                });
-
-                it('should call eth_feeHistory with newest block > latest', async function () {
-                    let latestBlock;
-                    const blocksAhead = 10;
-                    try {
-                        latestBlock = (await mirrorNode.get(`/blocks?limit=1&order=desc`, requestId)).blocks[0];
-                        const newestBlockNumberHex = ethers.utils.hexValue(latestBlock.number + blocksAhead);
-                        await relay.call('eth_feeHistory', ['0x1', newestBlockNumberHex, null], requestId);
-                    } catch (error) {
-                        Assertions.jsonRpcError(error, predefined.REQUEST_BEYOND_HEAD_BLOCK(latestBlock.number + blocksAhead, latestBlock.number));
-                    }
-                });
-
-                it('should call eth_feeHistory with zero block count', async function () {
-                    const res = await relay.call('eth_feeHistory', ['0x0', 'latest', null], requestId);
-
-                    expect(res.reward).to.not.exist;
-                    expect(res.baseFeePerGas).to.not.exist;
-                    expect(res.gasUsedRatio).to.equal(null);
-                    expect(res.oldestBlock).to.equal('0x0');
-                });
-            });
-        }
-
-        describe('eth_feeHistory', () => {
-            it('should call eth_feeHistory', async function () {
-                const res = await relay.call('eth_feeHistory', ['0x1', 'latest', null], requestId);
-                expect(res.baseFeePerGas).to.exist.to.be.an('Array');
-                expect(res.baseFeePerGas.length).to.be.gt(0);
-                expect(res.gasUsedRatio).to.exist.to.be.an('Array');
-                expect(res.gasUsedRatio.length).to.be.gt(0);
-                expect(res.oldestBlock).to.exist;
-                expect(Number(res.oldestBlock)).to.be.gt(0);
-            });
-        });
-
-        describe('Contract call reverts', () => {
-            let reverterContract, reverterEvmAddress;
-            const PURE_METHOD_CALL_DATA = '0xb2e0100c';
-            const VIEW_METHOD_CALL_DATA = '0x90e9b875';
-            const PAYABLE_METHOD_CALL_DATA = '0xd0efd7ef';
-            const PURE_METHOD_ERROR_DATA = '0x08c379a000000000000000000000000000000000000000000000000000000000000000200000000000000000000000000000000000000000000000000000000000000010526576657274526561736f6e5075726500000000000000000000000000000000';
-            const VIEW_METHOD_ERROR_DATA = '0x08c379a000000000000000000000000000000000000000000000000000000000000000200000000000000000000000000000000000000000000000000000000000000010526576657274526561736f6e5669657700000000000000000000000000000000';
-            const PAYABLE_METHOD_ERROR_DATA = '0x08c379a000000000000000000000000000000000000000000000000000000000000000200000000000000000000000000000000000000000000000000000000000000013526576657274526561736f6e50617961626c6500000000000000000000000000';
-            const PURE_METHOD_ERROR_MESSAGE = 'execution reverted: RevertReasonPure';
-            const VIEW_METHOD_ERROR_MESSAGE = 'execution reverted: RevertReasonView';
-
-            before(async () => {
-                reverterContract = await servicesNode.deployContract(reverterContractJson);
-                // Wait for creation to propagate
-                await mirrorNode.get(`/contracts/${reverterContract.contractId}`, requestId);
-                reverterEvmAddress = `0x${reverterContract.contractId.toSolidityAddress()}`;
-            });
-
-            it('Returns revert message for pure methods', async () => {
-                const callData = {
-                    from: '0x' + accounts[0].address,
-                    to: reverterEvmAddress,
-                    gas: EthImpl.numberTo0x(30000),
-                    data: PURE_METHOD_CALL_DATA
-                };
-
-                await relay.callFailing('eth_call', [callData, 'latest'], {
-                    code: -32008,
-                    message: PURE_METHOD_ERROR_MESSAGE,
-                    data: PURE_METHOD_ERROR_DATA
-                }, requestId);
-            });
-
-            it('Returns revert message for view methods', async () => {
-                const callData = {
-                    from: '0x' + accounts[0].address,
-                    to: reverterEvmAddress,
-                    gas: EthImpl.numberTo0x(30000),
-                    data: VIEW_METHOD_CALL_DATA
-                };
-
-                await relay.callFailing('eth_call', [callData, 'latest'], {
-                    code: -32008,
-                    message: VIEW_METHOD_ERROR_MESSAGE,
-                    data: VIEW_METHOD_ERROR_DATA
-                }, requestId);
-            });
-
-            it('Returns revert reason in receipt for payable methods', async () => {
-                const transaction = {
-                    value: ONE_TINYBAR,
-                    gasLimit: EthImpl.numberTo0x(30000),
-                    chainId: Number(CHAIN_ID),
-                    to: reverterEvmAddress,
-                    nonce: await relay.getAccountNonce('0x' + accounts[0].address, requestId),
-                    gasPrice: await relay.gasPrice(requestId),
-                    data: PAYABLE_METHOD_CALL_DATA
-                };
-                const signedTx = await accounts[0].wallet.signTransaction(transaction);
-                const transactionHash = await relay.call('eth_sendRawTransaction', [signedTx], requestId);
-
-                // Wait until receipt is available in mirror node
-                await mirrorNode.get(`/contracts/results/${transactionHash}`, requestId);
-
-                const receipt = await relay.call('eth_getTransactionReceipt', [transactionHash], requestId);
-                expect(receipt?.revertReason).to.exist;
-                expect(receipt.revertReason).to.eq(PAYABLE_METHOD_ERROR_DATA);
-            });
-
-            describe('eth_getTransactionByHash for reverted payable contract calls', async function() {
-                const payableMethodsData = [
-                    {
-                        data: '0xfe0a3dd7',
-                        method: 'revertWithNothing',
-                        message: '',
-                        errorData: '0x'
-                    },
-                    {
-                        data: '0x0323d234',
-                        method: 'revertWithString',
-                        message: 'Some revert message',
-                        errorData: '0x08c379a000000000000000000000000000000000000000000000000000000000000000200000000000000000000000000000000000000000000000000000000000000013536f6d6520726576657274206d65737361676500000000000000000000000000'
-                    },
-                    {
-                        data: '0x46fc4bb1',
-                        method: 'revertWithCustomError',
-                        message: '',
-                        errorData: '0x0bd3d39c'
-                    },
-                    {
-                        data: '0x33fe3fbd',
-                        method: 'revertWithPanic',
-                        message: '',
-                        errorData: '0x4e487b710000000000000000000000000000000000000000000000000000000000000012'
-                    }
-                ];
-                const hashes: any = [];
-
-                beforeEach(async () => {
-                    requestId = Utils.generateRequestId();
-                });
-
-                before(async function() {
-                    for (let i = 0; i < payableMethodsData.length; i++) {
-                        const transaction = {
-                            // value: ONE_TINYBAR,
-                            gasLimit: EthImpl.numberTo0x(30000),
-                            chainId: Number(CHAIN_ID),
-                            to: reverterEvmAddress,
-                            nonce: await relay.getAccountNonce('0x' + accounts[0].address, requestId),
-                            gasPrice: await relay.gasPrice(requestId),
-                            data: payableMethodsData[i].data
-                        };
-                        const signedTx = await accounts[0].wallet.signTransaction(transaction);
-                        const hash = await relay.call('eth_sendRawTransaction', [signedTx], requestId);
-                        hashes.push(hash);
-
-                        // Wait until receipt is available in mirror node
-                        await mirrorNode.get(`/contracts/results/${hash}`, requestId);
-                    }
-                });
-
-                for(let i = 0; i < payableMethodsData.length; i++) {
-                    it(`Payable method ${payableMethodsData[i].method} returns tx object`, async function () {
-                        const tx = await relay.call('eth_getTransactionByHash', [hashes[i]], requestId);
-                        expect(tx).to.exist;
-                        expect(tx.hash).to.exist;
-                        expect(tx.hash).to.eq(hashes[i]);
-                    });
-                }
-
-                describe('DEV_MODE = true', async function () {
-                    before(async () =>{
-                        process.env.DEV_MODE = 'true';
-                    });
-
-                    after(async () =>{
-                        process.env.DEV_MODE = 'false';
-                    });
-
-                    for(let i = 0; i < payableMethodsData.length; i++) {
-                        it(`Payable method ${payableMethodsData[i].method} throws an error`, async function () {
-                            await relay.callFailing('eth_getTransactionByHash', [hashes[i]], {
-                                code: -32008,
-                                message: payableMethodsData[i].message,
-                                data: payableMethodsData[i].errorData
-                            }, requestId);
-                        });
-                    }
-                });
-            });
-
-            describe('eth_call for reverted pure contract calls', async function() {
-                beforeEach(async () => {
-                    requestId = Utils.generateRequestId();
-                });
-
-                const pureMethodsData = [
-                    {
-                        data: '0x2dac842f',
-                        method: 'revertWithNothingPure',
-                        message: '',
-                        errorData: '0x'
-                    },
-                    {
-                        data: '0x8b153371',
-                        method: 'revertWithStringPure',
-                        message: 'Some revert message',
-                        errorData: '0x08c379a000000000000000000000000000000000000000000000000000000000000000200000000000000000000000000000000000000000000000000000000000000013536f6d6520726576657274206d65737361676500000000000000000000000000'
-                    },
-                    {
-                        data: '0x35314694',
-                        method: 'revertWithCustomErrorPure',
-                        message: '',
-                        errorData: '0x0bd3d39c'
-                    },
-                    {
-                        data: '0x83889056',
-                        method: 'revertWithPanicPure',
-                        message: '',
-                        errorData: '0x4e487b710000000000000000000000000000000000000000000000000000000000000012'
-                    }
-                ];
-
-                for (let i = 0; i < pureMethodsData.length; i++) {
-                    it(`Pure method ${pureMethodsData[i].method} returns tx receipt`, async function () {
-                        const callData = {
-                            from: '0x' + accounts[0].address,
-                            to: reverterEvmAddress,
-                            gas: EthImpl.numberTo0x(30000),
-                            data: pureMethodsData[i].data
-                        };
-
-                        await relay.callFailing('eth_call', [callData, 'latest'], {
-                            code: -32008,
-                            message: pureMethodsData[i].message,
-                            data: pureMethodsData[i].errorData
-                        }, requestId);
-                    });
-                }
             });
         });
     });
